% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.bam}
\alias{AlleleProfileR.bam}
\title{Load a .bam file as a data.frame}
\usage{
AlleleProfileR.bam(file)
}
\arguments{
\item{file}{Path to a .bam file}
}
\description{
This function will import a .bam file and stores it  as a data.frame.
}
\references{
Based on: https://gist.github.com/SamBuckberry/9914246
}
\author{
Arne Bruyneel
}
