% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.batch}
\alias{AlleleProfileR.batch}
\title{Start the analysis pipeline}
\usage{
AlleleProfileR.batch(config, cores = 1, subset = NULL,
  summarize = TRUE)
}
\arguments{
\item{config}{Configuration list, as generated by AlleleProfileR.setup.}

\item{cores}{Number of CPU cores to be used for processing.}

\item{subset}{List of the IDs of samples table and the IDs of the genes of interest in the genes table. If set to NULL, all samples will be analysed for all genes. Default is NULL.}

\item{summarize}{Generate ummary tables and exported as .csv files. Default is TRUE.}
}
\value{
This function does not return anything. All results are written to .csv files in the output folder.
}
\description{
This is the main function for processing the sequencing data to determine genetic variants.
}
\author{
Arne Bruyneel
}
