% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.batch.recompute}
\alias{AlleleProfileR.batch.recompute}
\title{Recompute summary output files}
\usage{
AlleleProfileR.batch.recompute(config, cutoff = 0, ignore.single = F,
  top = 0, subset = NULL)
}
\arguments{
\item{config}{Configuration list, as generated by AlleleProfileR.setup.}

\item{cutoff}{Percent value (0 to 1), representing the treshold read count for a certain allelic variant to be included in further analysis.}

\item{ignore.single}{Ingore alleles representing a single read pair? Default is FALSE.}

\item{top}{Only report the top x variants}

\item{subset}{List of the IDs of samples table and the IDs of the genes of interest in the genes table. If set to NULL, all samples will be analysed for all genes. Default is NULL.}
}
\value{
This function does not return anything. All results are written to .csv files in the output folder.
}
\description{
This function will recompute summary output files with defined cut-off value to exclude infrequent variants.
}
\author{
Arne Bruyneel
}
