% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.batch.summary}
\alias{AlleleProfileR.batch.summary}
\title{Summary function}
\usage{
AlleleProfileR.batch.summary(config, table = F, plot = T,
  param = "wt", subset = NULL)
}
\arguments{
\item{config}{Configuration list, as generated by AlleleProfileR.setup.}

\item{table}{Sumarized the results as a table?}

\item{plot}{Plot results?}

\item{param}{Parameter to plot. Default is wt.}

\item{subset}{List of the IDs of samples table and the IDs of the genes of interest in the genes table. If set to NULL, all samples will be analysed for all genes. Default is NULL.}
}
\value{
Summary table and/or heat map. The color of the heat map represents proportion, whereas the tile label indicates total number of alleles.
}
\description{
This function determines and plots the relative occurence of boolean features for all samples and all genes.
}
\author{
Arne Bruyneel
}
