% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.binreads}
\alias{AlleleProfileR.binreads}
\title{Bin reads based on genomic postition with respect to certain boundries}
\usage{
AlleleProfileR.binreads(bamtable, chr, pcrrange, range, purify = T)
}
\arguments{
\item{bamtable}{Data.frame containing the reads, as generated by AlleleProfileR.bam.}

\item{chr}{Chromosome}

\item{pcrrange}{Genomic range of that was pre-amplified by PCR before sequencing.}

\item{range}{Genomic range of interest, where variants will be called.}

\item{purify}{If set to TRUE, low qual reads will be discarded. Can be executed separately by AlleleProfileR.qc.simple.}
}
\value{
Bam data.frame with added columns 'bin', 'select', 'lower', and 'upper'.
}
\description{
This function will assign the reads within a certain PCR pre-amplification region to
different bins, depending on whether they do/do not/or partly overlap with the ROI.
A read will be classified as 'A' if it is within the PCR amplification range (pcrrange).
A read will be classified as 'B' or 'C' if it lies upstream and partly overlaps OR lies
downstream and partly overlaps with the region of interest (range). Finally, a read will be flagged as 'D'
if it fully overlaps with the ROI.
}
\author{
Arne Bruyneel
}
