% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.cryptic}
\alias{AlleleProfileR.cryptic}
\title{Identify cryptic coding sequences}
\usage{
AlleleProfileR.cryptic(seq, stopoptions = list("N", 0, NULL))
}
\arguments{
\item{seq}{Sequence, provide sequence including both the existing pre and post UTR regions}

\item{stopoptions}{List defining the stop conditions of the current exon: does it end in a stop codon,
or is it one of many exons. The first element of this list is the type of stopcodon (N for standard).
The second element is the shift in the codons, for example if the exon length is not a multiple of 3.
The third is the alternative stop codon position if it isn't a traditional stop codon (NULL for default).}
}
\value{
List of the cryptic proteins, and all start and stop codons respectively.
}
\description{
Determine whether there is a cryptic startcodon that may give rise to a cryptic 'protein'.
For example, if the endogenous startcodon is destroyed, but there is an alternative startcodon downstream.
}
\author{
Arne Bruyneel
}
