% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.determine.variants}
\alias{AlleleProfileR.determine.variants}
\title{Determine the allelic variants for a sample and gene}
\usage{
AlleleProfileR.determine.variants(bamtable, index, gene,
  cutoff.large = 25, ignore.snp = F, cut.range = 0,
  ignore.chimeric = F, chimeric.max.insertion.size = 50,
  suppress.messages = F)
}
\arguments{
\item{bamtable}{Data.frame containing the reads, as generated by AlleleProfileR.bam.}

\item{index}{Path to the .fa file containing the reference genome.}

\item{gene}{Gene information vector}

\item{cutoff.large}{Cutoff value for determining whether an indel is large or small. Default is 25.}

\item{ignore.snp}{Ignore SNPs? Default is FALSE.}

\item{cut.range}{Range around cut site for considering NHEJ. Default is 0, this setting will disregard cut range.}

\item{ignore.chimeric}{Ignore chimeric reads? Default is FALSE.}

\item{chimeric.max.insertion.size}{Max insertion size when inspecting chemeric alignments. Default is 50.}

\item{suppress.messages}{Suppress messages? Default is FALSE.}
}
\value{
Variants data.frame
}
\description{
This is the main worker function that parses through the reads in the bam file and determines their allele variants.
}
\author{
Arne Bruyneel
}
