% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.execute}
\alias{AlleleProfileR.execute}
\title{Wrapper execute function for the analysis}
\usage{
AlleleProfileR.execute(bamtable, samplestable, genestable, i, j, cutoff,
  index, ignore.snp, ignore.single, summarize = TRUE, cut.range,
  cutoff.large, ignore.chimeric, chimeric.max.insertion.size,
  suppress.messages = F)
}
\arguments{
\item{bamtable}{Data.frame containing the reads, as generated by AlleleProfileR.bam.}

\item{samplestable}{Samples table, inherited from config.}

\item{genestable}{Genes table, inherited from config.}

\item{i}{sample id}

\item{j}{gene id}

\item{cutoff}{Percent value (0 to 1), representing the treshold read count for a certain allelic variant to be included in further analysis.}

\item{index}{Path to the .fa file containing the reference genome.}

\item{ignore.snp}{Ignore SNPs? Default is FALSE.}

\item{ignore.single}{Ingore alleles representing a single read pair? Default is FALSE.}

\item{summarize}{Generate ummary tables and exported as .csv files. Default is TRUE.}

\item{cut.range}{Range around cut site for considering NHEJ. Default is 0, this setting will disregard cut range.}

\item{cutoff.large}{Cutoff value for determining whether an indel is large or small. Default is 25.}

\item{ignore.chimeric}{Ignore chimeric reads? Default is FALSE.}

\item{chimeric.max.insertion.size}{Max insertion size when inspecting chemeric alignments. Default is 50.}

\item{suppress.messages}{Suppress messages? Default is FALSE.}
}
\description{
This is a background function that will process a single sample for a single gene region.
}
\author{
Arne Bruyneel
}
