% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.getreads}
\alias{AlleleProfileR.getreads}
\title{Bin reads based on genomic postition with respect to certain boundries}
\usage{
AlleleProfileR.getreads(bamtable, chr, pcrrange, range, purify = T)
}
\arguments{
\item{bamtable}{Data.frame containing the reads, as generated by AlleleProfileR.bam.}

\item{chr}{Chromosome}

\item{pcrrange}{Genomic range of that was pre-amplified by PCR before sequencing.}

\item{range}{Genomic range of interest, where variants will be called.}

\item{purify}{If set to TRUE, low qual reads will be discarded. Can be executed separately by AlleleProfileR.qc.simple.}
}
\value{
Bam data.frame with added columns 'bin', 'select', 'lower', and 'upper'.
Only reads for which selected is TRUE will be retained (bin is D).
}
\description{
This function will only select reads that fully overlap with a genomic range (range),
and will ignore partly overlapping reads.
}
\author{
Arne Bruyneel
}
