% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bioinformatics.R
\name{AlleleProfileR.map}
\alias{AlleleProfileR.map}
\title{Align reads to the reference genome}
\usage{
AlleleProfileR.map(files, method = "bwa", index, subset = NULL)
}
\arguments{
\item{files}{Vector with files. Output of AlleleProfileR.read.folders}

\item{method}{Mapping method, default is BWA (BWA-MEM).}

\item{index}{Path to the .fa file containing the reference genome.}

\item{subset}{Vector with the indices of the entries that should be processed. Default is NULL, all entries will be analyzed.}
}
\description{
Align the reads to the reference using BWA. This external software needs to be installed manually and needs to be addressable from the terminal.
}
\references{
\itemize{
  \item Li, H. and Durbin, R. Fast and accurate long-read alignment with Burrows-Wheeler transform. Bioinformatics 2010;26(5):589-595.
}
}
\author{
Arne Bruyneel
}
