% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bioinformatics.R
\name{AlleleProfileR.merge}
\alias{AlleleProfileR.merge}
\title{Merge paired-end reads}
\usage{
AlleleProfileR.merge(files, method = "pear", params = "-v 10",
  subset = NULL)
}
\arguments{
\item{files}{Vector with files. Output of AlleleProfileR.read.folders}

\item{method}{Merge method, options are pear, fastq-join or flash. Default is pear (PEAR).}

\item{params}{String with additional parameters}

\item{subset}{Vector with the indices of the entries that should be processed. Default is NULL, all entries will be analyzed.}
}
\description{
Merge peared end reads using PEAR, FLASH or fastq-join. This external software needs to be installed manually and needs to be addressable from the terminal.
}
\references{
\itemize{
  \item Zhang, J., et al. PEAR: a fast and accurate Illumina Paired-End reAd mergeR. Bioinformatics 2014;30(5):614-620.
  \item Magoč, T. and Salzberg, S.L. FLASH: fast length adjustment of short reads to improve genome assemblies. Bioinformatics 2011;27(21):2957-2963.
  \item Aronesty, E. Comparison of Sequencing Utility Programs. The Open Bioinformatics Journal  2013;7:1-8.
}
}
\author{
Arne Bruyneel
}
