% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.parsecodingframe}
\alias{AlleleProfileR.parsecodingframe}
\title{Profile the codingframe sequence}
\usage{
AlleleProfileR.parsecodingframe(obj, gene, index, fs = F)
}
\arguments{
\item{obj}{Row in the BAM datatable, a sequencing read.}

\item{gene}{Gene information vector}

\item{index}{Path to the .fa file containing the reference genome.}

\item{fs}{If TRUE, a frameshift was already detected in previous analysis steps.}
}
\value{
List with elements: output_start, output_coding, output_fs, output_stop, output_pstop, length, output_wt, output_cryptic, seq, output_cryptic_start, output_cryptic_stop
}
\description{
This function profiles the coding sequence for variants
}
\author{
Arne Bruyneel
}
