% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.parseleadseq}
\alias{AlleleProfileR.parseleadseq}
\title{Parse the lead sequence in a chimeric pair}
\usage{
AlleleProfileR.parseleadseq(obj, gene, cutoff.large, index, cutrangelist)
}
\arguments{
\item{obj}{Row in the BAM datatable, a sequencing read.}

\item{gene}{Gene information vector}

\item{cutoff.large}{Cutoff value for determining whether an indel is large or small. Default is 25.}

\item{index}{Path to the .fa file containing the reference genome.}

\item{cutrangelist}{cutrangelist}
}
\value{
List with elements: output_fs, output_sm, output_lg, dels, list(ins_loc,ins_bps), startpos+atgpos-1, indelseq, output_utr, output_atg, output_inrange
}
\description{
This function processes the leading sequence in a chimeric pair.
}
\author{
Arne Bruyneel
}
