% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.parsetrailseq}
\alias{AlleleProfileR.parsetrailseq}
\title{Parse the trailing sequence in a chimeric pair}
\usage{
AlleleProfileR.parsetrailseq(obj, gene, cutoff.large, index, input_atg,
  cutrangelist)
}
\arguments{
\item{obj}{Row in the BAM datatable, a sequencing read.}

\item{gene}{Gene information vector}

\item{cutoff.large}{Cutoff value for determining whether an indel is large or small. Default is 25.}

\item{index}{Path to the .fa file containing the reference genome.}

\item{input_atg}{If TRUE, ATG is destroyed in the lead sequence}

\item{cutrangelist}{cutrangelist}
}
\value{
List with elements: output_fs, output_sm, output_lg, dels, list(ins_loc,ins_bps), indelseq, output_utr, output_inrange
}
\description{
Worker function to process the trailing sequence of a chimeric pair.
}
\author{
Arne Bruyneel
}
