% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.plot.alignment}
\alias{AlleleProfileR.plot.alignment}
\title{Alignment plot function}
\usage{
AlleleProfileR.plot.alignment(config, sample.id = NA, gene.id = NA,
  max.insertion.size = 20, ins.size = 5,
  legend.symbol.size = ins.size, legend.text.size = 6,
  plot.text.size = 2, axis.text.size = 8, alternate = NULL,
  alignmenttype = "local-global")
}
\arguments{
\item{config}{Configuration list, as generated by AlleleProfileR.setup.}

\item{sample.id}{Sample id, from samples object.}

\item{gene.id}{Gene id, from genes table.}

\item{max.insertion.size}{The maximum length of an insertion to be shown in the
legend. Larger insertions will be collapsed. Default is 20.}

\item{ins.size}{The size of the symbols representing insertions within the plot. Default is 5.}

\item{legend.symbol.size}{The size of the symbols indicating insertions in the legend. Default is ins.size.}

\item{legend.text.size}{The size of the legend labels. Default is 6.}

\item{plot.text.size}{The size of the text inside the plot. Default is 2.}

\item{axis.text.size}{The size of the axis labels. Default is 8.}

\item{alternate}{Alternate reference for labeling the HDR variant, as defined by AlleleProfileR.alternatereference.}

\item{alignmenttype}{Type for Biostrings::pairwiseAlignment operation to determine whether two alleles are identical. Default is "local-global".}
}
\value{
Alignment plot
}
\description{
THis function plots sequence alignments.
}
\references{
This function incorporates some of the plotting features of CrispRVariants (Helen Lindsay).
\itemize{
  \item Lindsay H, Burger A, Biyong B, Felker A, Hess C, Zaugg J, Chiavacci E, Anders C, Jinek M, Mosimann C, Robinson MD. CrispRVariants charts the mutation spectrum of genome engineering experiments. Nature Biotechnology 2016;34:701–702. doi: 10.1038/nbt.3628.
}
}
\author{
Arne Bruyneel
}
