% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.plot.mutationtypes}
\alias{AlleleProfileR.plot.mutationtypes}
\title{Plotting function}
\usage{
AlleleProfileR.plot.mutationtypes(config, sample.id = NA, gene.id = NA,
  alternate = NULL, plot = T, axis.text.size = 6, label = T,
  title = T)
}
\arguments{
\item{config}{Configuration list, as generated by AlleleProfileR.setup.}

\item{sample.id}{Sample id, from samples object.}

\item{gene.id}{Gene id, from genes table.}

\item{alternate}{Alternate reference for labeling the HDR variant, as defined by AlleleProfileR.alternatereference.}

\item{plot}{Plot results?}

\item{axis.text.size}{The size of the axis labels. Default is 8.}

\item{label}{Display value label above bars? Default is TRUE.}

\item{title}{Display title? Default is TRUE.}
}
\value{
AlleleProfileR.plot.mutationtypes
}
\description{
This is function will plot the proportion NHEJ, HDR and WT.
}
\author{
Arne Bruyneel
}
