% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.plot}
\alias{AlleleProfileR.plot}
\title{Plotting function}
\usage{
AlleleProfileR.plot(config, sample.id = NA, gene.id = NA,
  max.insertion.size = 20, ins.size = 5,
  legend.symbol.size = ins.size, legend.text.size = 6,
  plot.text.size = 2, axis.text.size = 8, alternate = NULL,
  alignmenttype = "local-global", type = "full", plotparam = c(2, 1))
}
\arguments{
\item{config}{Configuration list, as generated by AlleleProfileR.setup.}

\item{sample.id}{Sample id, from samples object.}

\item{gene.id}{Gene id, from genes table.}

\item{max.insertion.size}{The maximum length of an insertion to be shown in the
legend. Larger insertions will be collapsed. Default is 20.}

\item{ins.size}{The size of the symbols representing insertions within the plot. Default is 5.}

\item{legend.symbol.size}{The size of the symbols indicating insertions in the legend. Default is ins.size.}

\item{legend.text.size}{The size of the legend labels. Default is 6.}

\item{plot.text.size}{The size of the text inside the plot. Default is 2.}

\item{axis.text.size}{The size of the axis labels. Default is 8.}

\item{alternate}{Alternate reference for labeling the HDR variant, as defined by AlleleProfileR.alternatereference.}

\item{alignmenttype}{Type for Biostrings::pairwiseAlignment operation to determine whether two alleles are identical. Default is "local-global".}

\item{type}{full or count-only}

\item{plotparam}{barchart types}
}
\value{
AlleleProfileR.plot
}
\description{
This is the main plotting function of this R-package.
}
\author{
Arne Bruyneel
}
