% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.plot.readdept}
\alias{AlleleProfileR.plot.readdept}
\title{Histogram of the read-depth}
\usage{
AlleleProfileR.plot.readdept(config, sample.id = NA, gene.id = NA,
  plot = TRUE, table = FALSE)
}
\arguments{
\item{config}{Configuration list, as generated by AlleleProfileR.setup.}

\item{sample.id}{Sample id, from samples object.}

\item{gene.id}{Gene id, from genes table.}

\item{plot}{Plot results?}

\item{table}{Sumarized the results as a table?}
}
\value{
Histogram plot and data table
}
\description{
This function determines the read-depth of a sample and gene combination.
}
\author{
Arne Bruyneel
}
