% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.plot.seqlogo}
\alias{AlleleProfileR.plot.seqlogo}
\title{Seqlogo plot function}
\usage{
AlleleProfileR.plot.seqlogo(config, sample.id = NA, gene.id = NA,
  customrange = c(NA, NA), plot = F, axis.text.size = 8)
}
\arguments{
\item{config}{Configuration list, as generated by AlleleProfileR.setup.}

\item{sample.id}{Sample id, from samples object.}

\item{gene.id}{Gene id, from genes table.}

\item{customrange}{Vector with start and stop position of the region to be analyzed/plotted.}

\item{plot}{Plot results?}

\item{axis.text.size}{The size of the axis labels. Default is 8.}
}
\value{
plot ggseqlogo
}
\description{
Plot the efficacy of base-editing as a sequence logo.
}
\references{
\itemize{
  \item Wagih, O. ggseqlogo: a versatile R package for drawing sequence logos. Bioinformatics 2017;33(22):3645-3647.
}
}
\author{
Arne Bruyneel
}
