% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bioinformatics.R
\name{AlleleProfileR.preprocess}
\alias{AlleleProfileR.preprocess}
\title{Preprocess fastq.gz files}
\usage{
AlleleProfileR.preprocess(files, index, method.qc = "fastp",
  params.qc = "", method.merge = "pear", params.merge = "-v 10",
  method.map = "bwa", subset = NULL)
}
\arguments{
\item{files}{Vector with files. Output of AlleleProfileR.read.folders}

\item{index}{Path to the .fa file containing the reference genome.}

\item{method.qc}{Method for conducting QC on raw reads. Currently only fastp is implemented. Use 'none' to skip this step.}

\item{params.qc}{Additional parameters for conducting QC on raw reads.}

\item{method.merge}{Method for merging paired reads. Currently only pear and fastq-join are implemented.}

\item{params.merge}{Additional parameters for merging paired reads}

\item{method.map}{Method for mapping reads to reference. Currently only BWA is implemented.}

\item{subset}{Vector with the indices of the entries that should be processed. Default is NULL, all entries will be analyzed.}
}
\description{
This function executes QC (fastp), merges paired reads (pear), and maps the reads to the reference (BWA). This external software needs to be installed manually and needs to be addressable from the terminal.
}
\author{
Arne Bruyneel
}
