% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bioinformatics.R
\name{AlleleProfileR.qc}
\alias{AlleleProfileR.qc}
\title{QC reads using external software}
\usage{
AlleleProfileR.qc(files, method = "fastp", params = "",
  subset = NULL)
}
\arguments{
\item{files}{Vector with files. Output of AlleleProfileR.read.folders}

\item{method}{QC method. Default is fastp.}

\item{params}{String with additional parameters}

\item{subset}{Vector with the indices of the entries that should be processed. Default is NULL, all entries will be analyzed.}
}
\description{
QC reads using external software. Currently only fastp is available. This external software needs to be installed manually and needs to be addressable from the terminal.
}
\references{
\itemize{
  \item Shifu Chen, Yanqing Zhou, Yaru Chen, Jia Gu; fastp: an ultra-fast all-in-one FASTQ preprocessor, Bioinformatics, Volume 34, Issue 17, 1 September 2018, Pages i884–i890, https://doi.org/10.1093/bioinformatics/bty560
}
}
\author{
Arne Bruyneel
}
