% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.read.folders}
\alias{AlleleProfileR.read.folders}
\title{Read folders for .fastq or .bam files}
\usage{
AlleleProfileR.read.folders(type = "fastq")
}
\arguments{
\item{type}{Input data type in the files/input folder, use either 'fastq' or 'bam'. If QC processing was conducted using AlleleProfileR, use fastq-clean.}
}
\value{
Sample table, listing the fastq or bam files for each sample.
}
\description{
This function will scan the subfolders of files/input folder for .fastq.gz or .bam files,
or the files/output folder for .bam files.
To generate .bam files from .fastq files, the input folders should be scanned using the type = "fastq" parameter.
Next, to conduct the analysis on .bam files that were generated in this step, use type = "bam".
Please refer to the tutorial for more details.
}
\author{
Arne Bruyneel
}
