% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.sample.distribution.boolean}
\alias{AlleleProfileR.sample.distribution.boolean}
\title{Allelelic variant distribution summary and plotting function for boolean outcome variables}
\usage{
AlleleProfileR.sample.distribution.boolean(pipe_config, sample.id = NA,
  gene.id = NA, plot = TRUE, plotparam = c("wt", "fs", "snp", "atg",
  "coding", "stop", "pstop", "sm", "lg", "utr", "cryptic", "error"),
  alternate = NULL, title = TRUE, axis.text.size = 8,
  legend.text.size = 6, display_other = T, ylabels = TRUE,
  order = NULL)
}
\arguments{
\item{pipe_config}{congifuration object}

\item{sample.id}{Sample id, from samples object.}

\item{gene.id}{Gene id, from genes table.}

\item{plot}{Plot results?}

\item{plotparam}{names of the parameters to plot}

\item{alternate}{Alternate reference for labeling the HDR variant, as defined by AlleleProfileR.alternatereference.}

\item{title}{If TRUE, a title will be included in the graph.}

\item{axis.text.size}{The size of the axis labels. Default is 8.}

\item{legend.text.size}{The size of the legend labels. Default is 6.}

\item{display_other}{Display the 'other' allele. Default is TRUE.}

\item{ylabels}{Print ylabels. Default is TRUE.}

\item{order}{ordering variable}
}
\value{
Table and/or plot.
}
\description{
This function determines summary statistics and plots
heatmaps of the allelic variants.
}
\author{
Arne Bruyneel
}
