% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.sample.distribution}
\alias{AlleleProfileR.sample.distribution}
\title{Allelelic variant distribution summary and plotting function}
\usage{
AlleleProfileR.sample.distribution(pipe_config, sample.id = NA,
  gene.id = NA, plot = TRUE, plotparam = 1, alternate = NULL,
  title = TRUE, axis.text.size = 8, legend.text.size = 6,
  display_other = TRUE, ylabels = TRUE, order = NULL,
  emptyrow = TRUE, addwt = TRUE)
}
\arguments{
\item{pipe_config}{congifuration object}

\item{sample.id}{Sample id, from samples object.}

\item{gene.id}{Gene id, from genes table.}

\item{plot}{Plot results?}

\item{plotparam}{If 1, the coding frame length will be plotted. If 2, the read count will be plotted.}

\item{alternate}{Alternate reference for labeling the HDR variant, as defined by AlleleProfileR.alternatereference.}

\item{title}{If TRUE, a title will be included in the graph.}

\item{axis.text.size}{The size of the axis labels. Default is 8.}

\item{legend.text.size}{The size of the legend labels. Default is 6.}

\item{display_other}{Display the 'other' allele. Default is TRUE.}

\item{ylabels}{Print ylabels. Default is TRUE.}

\item{order}{ordering variable}

\item{emptyrow}{Add emptyrow bool. Default is TRUE.}

\item{addwt}{Add addwt bool. Default is TRUE.}
}
\value{
Table and/or plot.
}
\description{
This function determines summary statistics and plots
barcharts of the allelic variants, for either coding sequence length or read counts.
}
\author{
Arne Bruyneel
}
