% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.setup}
\alias{AlleleProfileR.setup}
\title{Set the configuration parameters for executing an analysis}
\usage{
AlleleProfileR.setup(samples, genes, index, cutoff = 0.02,
  ignore.snp = F, ignore.single = F, ignore.chimeric = F,
  cut.range = 0, alternate = NULL, cutoff.large = 25,
  chimeric.max.insertion.size = 50, suppress.messages = F)
}
\arguments{
\item{samples}{Table with all samples entries, as generated by AlleleProfileR.read.folders.}

\item{genes}{Path to the .csv file containing all gene entries.}

\item{index}{Path to the .fa file containing the reference genome.}

\item{cutoff}{Percent value (0 to 1), representing the treshold read count for a certain allelic variant to be included in further analysis.}

\item{ignore.snp}{Ignore SNPs? Default is FALSE.}

\item{ignore.single}{Ingore alleles representing a single read pair? Default is FALSE.}

\item{ignore.chimeric}{Ignore chimeric reads? Default is FALSE.}

\item{cut.range}{Range around cut site for considering NHEJ. Default is 0, this setting will disregard cut range.}

\item{alternate}{List of alternate objects for all samples (in order), NULL if not used.}

\item{cutoff.large}{Cutoff value for determining whether an indel is large or small. Default is 25.}

\item{chimeric.max.insertion.size}{Max insertion size when inspecting chemeric alignments. Default is 50.}

\item{suppress.messages}{Suppress messages? Default is FALSE.}
}
\value{
list
}
\description{
This function will generate a configuration list based on the provided arguments. This list will have the provide in further analysis steps.
}
\author{
Arne Bruyneel
}
