% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bioinformatics.R
\name{AlleleProfileR.simulate}
\alias{AlleleProfileR.simulate}
\title{Simulate paired-end reads from fasta files}
\usage{
AlleleProfileR.simulate(input.folder, input.file,
  params = "-e 0 -d 250 -N 50000 -1 150 -2 150 -r 0 -R 0 -X 0")
}
\arguments{
\item{input.folder}{Path to input fasta file folder}

\item{input.file}{Vector with one or more .fas file names}

\item{params}{Parameters for wgsim, default is: -e 0 -d 250 -N 50000 -1 150 -2 150 -r 0 -R 0 -X 0}
}
\description{
This function samples reads from one or more fasta file using wgsim. This external software needs to be installed manually and needs to be addressable from the terminal.
No data is returned: fastq.gz files are generated and deposited in the input.folder directory.
}
\references{
\itemize{
  \item https://github.com/lh3/wgsim
}
}
