% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{AlleleProfileR.summary.profiled.alleles}
\alias{AlleleProfileR.summary.profiled.alleles}
\title{Summary statistics of the profiled alleles}
\usage{
AlleleProfileR.summary.profiled.alleles(profiledvariants, cutoff = 0,
  ignore.single = F, top = 0, alternateallele = NULL)
}
\arguments{
\item{profiledvariants}{Variants data.frame, as defined by AlleleProfileR.determine.variants.}

\item{cutoff}{Percent value (0 to 1), representing the treshold read count for a certain allelic variant to be included in further analysis.}

\item{ignore.single}{Ingore alleles representing a single read pair? Default is FALSE.}

\item{top}{Only report the top x variants}

\item{alternateallele}{Alternate allele name}
}
\value{
Detailed summary table. Allele: allele name, count: read counts, qname: identifier of a representative read, wt: wild type coding sequence,
snp: no of SNPs, atg: ATG destroyed, coding: coding for a protein, coding length: length of the exon, fs: frame shift present,
stop: stop destroyed, pstop: premature stop inserted, sm: small indel, lg: large indel, utr: indel in UTR region,
cryptic: cryptic coding frame present, error: errors in analysis, seq: read sequence, cdsseq: coding frame sequence,
pos: start position, cigar: cigar string, indelseq: indel sequences.
}
\description{
This function generates summary statistics of the profiled alleles of a sample.
}
\author{
Arne Bruyneel
}
